CR-T14: prevent FMA fusion on the synthesis path (INV-DETERMINISM, D-011)
Three expressions had the form x + y*z, which the Go spec lets an implementation
fuse into a single FMA with one rounding. The gc compiler does fuse on
arm64/ppc64/s390x and does not on amd64, so the same fixture and seed would
synthesize different data on Apple Silicon than in amd64 CI.

Measured rather than assumed -- and hand-picked values showed nothing, which
would have closed this as theoretical. Searching 2,000,000 draws at realistic
magnitudes, fusing changes 14.15% of float results and 0.0725% of the int64
values that reach the SQL: about one row in 1,380.

All three sites now wrap the product in float64(), the remedy the spec itself
specifies.

Enforced by a golden hydrate digest run on both amd64 and arm64 in CI, plus a
test that fails if fusion ever stops changing an emitted value -- so the
conversions cannot quietly become cargo.

Mutation-tested by simulating the fusing backend with math.FMA: different
digest, proving the golden would catch a real cross-arch divergence.

Not yet observed on real arm64 hardware; see D-011.

+++ internal/hydrate/engine.go
@@ -176,7 +176,20 @@ func sampleHistogram(h *fixture.Histogram, r *rng) (any, bool) {
 		lo, hi = hi, lo
 	}
 	span := hi - lo
-	v := lo + r.float64()*span
+	// The float64() conversion is LOAD-BEARING, not decoration. The Go spec lets
+	// an implementation fuse a multiply and an add into a single FMA with only
+	// one rounding, and the compiler DOES fuse this shape on arm64, ppc64 and
+	// s390x while amd64 does not — so the same fixture and seed would synthesize
+	// different values on an Apple Silicon laptop than in amd64 CI, breaking
+	// INV-DETERMINISM's "byte-identical on ANY platform".
+	//
+	// The spec also gives the remedy: "An explicit floating-point type conversion
+	// rounds to the precision of the target type, preventing fusion that would
+	// discard that rounding." Measured over 2M randomly drawn (lo, r, span) at
+	// realistic fixture magnitudes, fusing changes 14.15% of the float results
+	// and 0.07% of the int64 values that actually reach the SQL — roughly one row
+	// in 1,380. Do not "simplify" this back.
+	v := lo + float64(r.float64()*span)
 	return int64(v), true
 }
